pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-clinicalops"
version = "0.1.0"
description = "ClinicalOps — Hospital clinical workflow environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
server = "clinicalops.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["clinicalops", "clinicalops.server"]
package-dir = { "clinicalops" = ".", "clinicalops.server" = "server" }
